refactor(frontend): decompose AudioPlayer, PlaylistDetailView, ConfigView (Phase C)
Splits the three remaining oversized view components into focused
subcomponents. Behavior is unchanged and all 168 frontend tests still pass.

- AudioPlayer.tsx (895 → 511) extracts repeated SVGs to audioPlayer/icons.tsx,
  the artwork+lyrics overlay to PlayerArtwork, the title/artist/album block to
  PlayerTrackInfo, the duplicated mobile control panel to
  PlayerMobileOptionsPanel, and the empty state to PlayerEmpty. The hook-level
  pieces the prior roadmap mentioned (useMediaSession, useTranscodeFallback)
  already live in useAudioPlayback.
- PlaylistDetailView.tsx (837 → 457) extracts PlaylistCover (with the mosaic
  helper), SortableTrackItem, the DnD-wrapped PlaylistEditableTrackList,
  PlaylistActions (the play/shuffle/edit/delete/save/cancel button row), and
  CollaboratorsField.
- ConfigView.tsx (745 → 71) becomes a thin switcher that delegates to
  IndexOpsSection or FilesSection. The bulk-delete and bulk-edit modals each
  live in their own files under config/.
